+++ api/mapping/test_views.py
@@ -1,7 +1,7 @@
 import os
 from django.test import TestCase
 from django.contrib.auth import get_user_model
-from rest_framework.test import APIRequestFactory, force_authenticate
+from rest_framework.test import APIRequestFactory, APIClient, force_authenticate
 from rest_framework.authtoken.models import Token
 from .views import DatasetListView, ScanReportListViewSet
 from .models import Project, Dataset, ScanReport, VisibilityChoices
@@ -53,7 +53,7 @@ class TestDatasetListView(TestCase):
 
     def test_dataset_returns(self):
         # Make the request for Datasets
-        request = self.factory.get(f"api/datasets/")
+        request = self.factory.get(f"/api/datasets/")
         # Add user1 to the request; this is not automatic
         request.user = self.user1
         # Authenticate the user1
@@ -102,7 +102,7 @@ class TestDatasetListView(TestCase):
     def test_dataset_filtering(self):
         # Make the request for the public_dataset1
         request = self.factory.get(
-            f"api/datasets/", {"id__in": self.public_dataset1.id}
+            f"/api/datasets/", {"id__in": self.public_dataset1.id}
         )
         # Add user1 to the request; this is not automatic
         request.user = self.user1
@@ -122,7 +122,7 @@ class TestDatasetListView(TestCase):
 
         # Make the request for the public_dataset3
         request = self.factory.get(
-            f"api/datasets/", {"id__in": self.public_dataset3.id}
+            f"/api/datasets/", {"id__in": self.public_dataset3.id}
         )
         # Add user1 to the request; this is not automatic
         request.user = self.user1
@@ -142,7 +142,7 @@ class TestDatasetListView(TestCase):
         User = get_user_model()
         az_user = User.objects.get(username=os.getenv("AZ_FUNCTION_USER"))
         # Make the request for the Dataset
-        request = self.factory.get(f"api/datasets/")
+        request = self.factory.get(f"/api/datasets/")
         # Add the user to the request; this is not automatic
         request.user = az_user
         # Authenticate az_user
@@ -155,6 +155,126 @@ class TestDatasetListView(TestCase):
         response_data = self.view(request).data
         # Assert az_user can see all datasets
         self.assertEqual(len(response_data), Dataset.objects.all().count())
+
+
+class TestDatasetUpdateView(TestCase):
+    def setUp(self):
+        User = get_user_model()
+        # Set up users
+        self.admin_user = User.objects.create(
+            username="gandalf", password="hjfiwejfiwef"
+        )
+        Token.objects.create(user=self.admin_user)
+        self.non_admin_user = User.objects.create(
+            username="aragorn", password="djfoiejwiofjoiewf"
+        )
+        Token.objects.create(user=self.non_admin_user)
+        self.non_project_user = User.objects.create(
+            username="bilbo", password="djfoiejwiofjoiewf"
+        )
+        Token.objects.create(user=self.non_project_user)
+
+        # Set up Project
+        self.project = Project.objects.create(name="The Fellowship of the Ring")
+        self.project.members.add(self.admin_user, self.non_admin_user)
+
+        # Set up Dataset
+        self.dataset = Dataset.objects.create(
+            name="The Heights of Hobbits", visibility=VisibilityChoices.PUBLIC
+        )
+        self.dataset.admins.add(self.admin_user)
+        self.project.datasets.add(self.dataset)
+
+        # Request factory for setting up requests
+        self.client = APIClient()
+
+    def test_update_returns(self):
+        # Authenticate admin user
+        self.client.force_authenticate(self.admin_user)
+        #  Make the request
+        response = self.client.patch(
+            f"/api/datasets/update/{self.dataset.id}", data={"name": "The Two Towers"}
+        )
+        response_data = response.data
+        # Ensure admin user can update Dataset
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response_data.get("name"), "The Two Towers")
+
+    def test_non_admin_member_forbidden(self):
+        # Authenticate non admin user
+        self.client.force_authenticate(self.non_admin_user)
+        #  Make the request
+        response = self.client.patch(
+            f"/api/datasets/update/{self.dataset.id}", data={"name": "The Two Towers"}
+        )
+        # Ensure non admin user is Forbidden
+        self.assertEqual(response.status_code, 403)
+
+    def test_non_project_member_forbidden(self):
+        # Authenticate non project user
+        self.client.force_authenticate(self.non_project_user)
+        #  Make the request
+        response = self.client.patch(
+            f"/api/datasets/update/{self.dataset.id}", data={"name": "The Two Towers"}
+        )
+        # Ensure non project user is Forbidden
+        self.assertEqual(response.status_code, 403)
+
+
+class TestDatasetDeleteView(TestCase):
+    def setUp(self):
+        User = get_user_model()
+        # Set up users
+        self.admin_user = User.objects.create(
+            username="gandalf", password="hjfiwejfiwef"
+        )
+        Token.objects.create(user=self.admin_user)
+        self.non_admin_user = User.objects.create(
+            username="aragorn", password="djfoiejwiofjoiewf"
+        )
+        Token.objects.create(user=self.non_admin_user)
+        self.non_project_user = User.objects.create(
+            username="bilbo", password="djfoiejwiofjoiewf"
+        )
+        Token.objects.create(user=self.non_project_user)
+
+        # Set up Project
+        self.project = Project.objects.create(name="The Fellowship of the Ring")
+        self.project.members.add(self.admin_user, self.non_admin_user)
+
+        # Set up Dataset
+        self.dataset = Dataset.objects.create(
+            name="The Heights of Hobbits", visibility=VisibilityChoices.PUBLIC
+        )
+        self.dataset.admins.add(self.admin_user)
+        self.project.datasets.add(self.dataset)
+
+        # Request factory for setting up requests
+        self.client = APIClient()
+
+    def test_update_returns(self):
+        # Authenticate admin user
+        self.client.force_authenticate(self.admin_user)
+        #  Make the request
+        response = self.client.delete(f"/api/datasets/delete/{self.dataset.id}")
+        # Ensure admin user can delete Dataset
+        self.assertEqual(response.status_code, 204)
+
+    def test_non_admin_member_forbidden(self):
+        # Authenticate non admin user
+        self.client.force_authenticate(self.non_admin_user)
+        #  Make the request
+        response = self.client.delete(f"/api/datasets/delete/{self.dataset.id}")
+        # Ensure non admin user is Forbidden
+        self.assertEqual(response.status_code, 403)
+
+    def test_non_project_member_forbidden(self):
+        # Authenticate non project user
+        self.client.force_authenticate(self.non_project_user)
+        #  Make the request
+        response = self.client.delete(f"/api/datasets/delete/{self.dataset.id}")
+        # Ensure non project user is Forbidden
+        self.assertEqual(response.status_code, 403)
 
 
 class TestScanScanReportListViewset(TestCase):